Moved and added tests for gnps format
